+++ src/compas_eve/mqtt/__init__.py
@@ -17,11 +17,6 @@ Classes
 
 """
 
-import sys
-
-if sys.platform == "cli":
-    from .mqtt_cli import MqttTransport
-else:
-    from .mqtt_paho import MqttTransport
+from .mqtt_paho import MqttTransport
 
 __all__ = ["MqttTransport"]